pyproject.toml
[tool.poetry]
name = "lexicon"
version = "0.1.0"
description = "FFXIV Equipment Auto-labeling Tool"
authors = ["AetherSight"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/AetherSight/Lexicon"
repository = "https://github.com/AetherSight/Lexicon"
keywords = ["ffxiv", "equipment", "labeling", "ai", "vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = ">=3.12"
openai = ">=1.0.0"
ollama = ">=0.1.0"
pandas = ">=2.0.0"
tqdm = ">=4.66.0"
pillow = ">=10.0.0"
fastapi = ">=0.104.0"
uvicorn = ">=0.24.0"
python-multipart = ">=0.0.6"
python-dotenv = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
ruff = ">=0.1.0"

[tool.poetry.scripts]
lexicon = "lexicon.train:train_main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
